Changed using float for purity do double due to rounding issues

+++ src/main/java/com/skcraft/smes/item/ItemPurifiable.java
@@ -23,6 +23,7 @@ public class ItemPurifiable extends ItemBase implements IToolTipProvider {
     @Override
     public void provideTooltip(ItemStack itemStack, EntityPlayer player, List<String> toolTip) {
         String tooltip = itemStack.getUnlocalizedName().replace("item." + SMES.PREFIX, "") + ".desc";
+        
         toolTip.add(String.format(StringUtils.translate(tooltip, true), 
                     itemStack.getTagCompound().getDouble(SMES.PREFIX + "purity")));
     }
@@ -31,22 +32,22 @@ public class ItemPurifiable extends ItemBase implements IToolTipProvider {
     @SideOnly(Side.CLIENT)
     public void getSubItems(Item item, CreativeTabs creativeTabs, List list) {
         // First NBT version with 55.0000000% purity
-        list.add(createSubType(item, 0.550000000f));
+        list.add(createSubType(item, 0.550000000));
         
         // Second NBT version with 80.0000000% purity
-        list.add(createSubType(item, 0.800000000f));
+        list.add(createSubType(item, 0.800000000));
         
         // Third NBT version with 95.0000000% purity
-        list.add(createSubType(item, 0.950000000f));
+        list.add(createSubType(item, 0.950000000));
     }
     
-    private ItemStack createSubType(Item item, float purity) {
+    private ItemStack createSubType(Item item, double purity) {
         // Prepare itemStack and NBTTagCompound
         NBTTagCompound nbttagcompound = new NBTTagCompound();
         StringBuilder sb = new StringBuilder(SMES.PREFIX).append("purity");
         ItemStack itemStack = new ItemStack(item, 1, 0);
         
-        nbttagcompound.setFloat(sb.toString(), purity);
+        nbttagcompound.setDouble(sb.toString(), purity);
         itemStack.setTagCompound(nbttagcompound);
         
         return itemStack;
